+++ crates/ruma-events/src/lib.rs
@@ -189,7 +189,11 @@ pub mod room_key;
 #[cfg(feature = "unstable-msc4268")]
 pub mod room_key_bundle;
 pub mod room_key_request;
-#[cfg(feature = "unstable-msc4310")]
+#[cfg(any(
+    feature = "unstable-msc4075",
+    feature = "unstable-msc4310",
+    feature = "unstable-msc4143"
+))]
 pub mod rtc;
 pub mod secret;
 pub mod secret_storage;