+++ crates/ruma-events/src/rtc.rs
@@ -1,6 +1,16 @@
 //! Modules for events in the `m.rtc` namespace.
 
+#[cfg(feature = "unstable-msc4143")]
+pub mod application;
 #[cfg(feature = "unstable-msc4310")]
 pub mod decline;
+#[cfg(feature = "unstable-msc4143")]
+pub mod encryption_key;
+#[cfg(feature = "unstable-msc4143")]
+pub mod member;
 #[cfg(feature = "unstable-msc4075")]
 pub mod notification;
+#[cfg(feature = "unstable-msc4143")]
+pub mod slot;
+#[cfg(feature = "unstable-msc4143")]
+pub mod transport;